// api/src/swapi/interfaces/IPeople.ts
export interface IPeopleFull {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
  created: string;
  edited: string;
  url: string;
}

export interface IPeople
  extends Pick<IPeopleFull, 'name' | 'height' | 'mass' | 'gender'> {}